+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     - Add Phasor Octature module.
     - Add Phasor Timetable module.
     - Add Phasor Stutter module.
+    - Add Phasor Shift module.
 - Add Gate Delay module.
 - Add Thomas chaotic attractor to 1-Op Chaos (https://sprott.physics.wisc.edu/chaos/symmetry.htm).
 - Make Mode knobs on 1-Op Chaos, 2-Op Chaos, Chaotic Attactors, and Clocked Noise snap to values and print the current mode.